proverka

+++ controllers/cards.js
@@ -32,17 +32,23 @@ const createCard = (req, res, next) => {
 };
 
 const deleteCard = (req, res, next) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => {
-      if (!card) {
-        throw new NotFoundError('Информация по карточке не найдена.');
+  Card.findById(req.params.cardId)
+    .then((reqCard) => {
+      if (!reqCard) {
+        throw new NotFoundError('Некорректный Id');
       }
-      if (card.owner.toString() !== req.user._id) {
-        throw new ForbiddenError('Нет прав на удаление.');
+      if (reqCard.owner.toString() === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => {
+            res.send({ data: card });
+          });
+      } else {
+        next(new ForbiddenError('Это не ваша карточка'));
       }
-      return res.send(card);
     })
-    .catch(next);
+    .catch((err) => {
+      next(err);
+    });
 };
 
 const likeCard = (req, res, next) => {
